requirements.txt
discord.py==2.3.2
pytz==2025.2
python-dotenv==1.1.0
pyserial==3.5
psutil==7.0.0
google-generativeai==0.8.5